fix edge case

+++ bit_manipulation/2-get_bit.c
@@ -7,7 +7,7 @@
  */
 int get_bit(unsigned long int n, unsigned int index)
 {
-	int bit;
+	unsigned long int bit;
 	
 	if (index == 1)
 	{
